+++ lib/keyboard.dart
@@ -1,0 +1,6 @@
+import 'package:flutter/services.dart';
+import './game.dart';
+
+mixin KeyboardEvents on Game {
+  void onKeyEvent(RawKeyEvent event);
+}