+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [next] 
 - Adding Component#onDestroy
+- Adding Keyboard events API
 
 ## 0.18.2
 - Add loop for AnimationComponent.sequenced() (thanks @wenxiangjiang)